requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
anyio==3.7.1
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.1.0
Brotli==1.0.9
cachetools==5.3.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.3.1
click==8.1.7
colorama==0.4.6
cryptography==41.0.3
cycler==0.12.1
dataclasses-json==0.5.7
docker-pycreds==0.4.0
et-xmlfile==1.1.0
exceptiongroup==1.1.3
Flask==2.2.2
fonttools==4.43.1
frozenlist==1.3.3
gitdb==4.0.11
GitPython==3.1.40
google-api-core==2.12.0
google-auth==2.23.3
googleapis-common-protos==1.61.0
greenlet==2.0.1
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.0.1
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
kiwisolver==1.4.5
langchain==0.0.322
langsmith==0.0.51
MarkupSafe==2.1.1
marshmallow==3.20.1
marshmallow-enum==1.5.1
matplotlib==3.5.3
mkl-service==2.4.0
multidict==6.0.2
munkres==1.1.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.22.4
openai==0.28.1
openapi-schema-pydantic==1.2.4
openpyxl==3.1.2
packaging==23.2
pandas==2.1.1
pandas-stubs==1.2.0.62
pathtools==0.1.2
Pillow==9.4.0
pip==23.3
plotly==5.17.0
protobuf==3.20.3
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.12
pyOpenSSL==23.2.0
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pyu2f==0.1.5
PyYAML==6.0
requests==2.31.0
rsa==4.9
scikit-learn==1.3.2
scipy==1.11.3
sentry-sdk==1.32.0
setproctitle==1.3.3
setuptools==68.0.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
SQLAlchemy==1.4.13
stringcase==1.2.0
tenacity==8.2.3
threadpoolctl==3.2.0
tqdm==4.66.1
typing_extensions==4.8.0
typing-inspect==0.9.0
tzdata==2023.3
unicodedata2==15.1.0
urllib3==2.0.7
wandb==0.15.12
Werkzeug==2.2.3
wheel==0.41.2
win-inet-pton==1.1.0
yarl==1.7.2
zipp==3.11.0
